Convert simulation stack row test to bun:test
- Replace the micro-should test entrypoint with Bun test cases
- Add coverage for message-only rows and mixed pending row order
